refactor: Move UpdateM1 function to blockchain_viction.go and add new backend_viction.go file for Viction-specific functionalities
